refactor: converters, extractors, mp3_to_wav and app
